fix(admin-api) declarative config yaml loading
This PR adds a way to load declarative config as an inline YAML file.

Previously it was possible only as a YAML string inside the the `config`
JSON payload. Now JSON and YAML have parity of ways to being loaded.

In addition to that fix, the reading algorithm has been changed so that
JSON is always tried first, then YAML, then Lua (when enabled).

This fixes a previous problem in which the YAML parser was used
erroneously to parse JSON configs.

This last fix is what made us make those slight changes in the tests.
Things like trailing commas were accepted by the YAML parser but are not
valid JSON.